use 64bit ints for data in count then move

+++ assignment1/count-then-move.cpp
@@ -30,7 +30,7 @@ const string PROGRAM_NAME = "count-then-move";
 const int MAX_BUCKETS = 256;
 
 // Add some computation to better demonstrate multi-core benefits
-void do_computation(tuple<int32_t, int32_t> &item)
+void do_computation(tuple<int64_t, int64_t> &item)
 {
   // Simple but non-trivial computation
   for (int i = 0; i < 50; i++)
@@ -44,12 +44,12 @@ void do_computation(tuple<int32_t, int32_t> &item)
  * @param n The number to get the data for.
  * @return The data for the number.
  */
-vector<tuple<int32_t, int32_t>> get_data_given_n(int n)
+vector<tuple<int64_t, int64_t>> get_data_given_n(int n)
 {
-  vector<tuple<int32_t, int32_t>> data(n);
-  for (int32_t i = 0; i < n; i++)
+  vector<tuple<int64_t, int64_t>> data(n);
+  for (int64_t i = 0; i < n; i++)
   {
-    data[i] = tuple<int32_t, int32_t>(i + 1, i + 1);
+    data[i] = tuple<int64_t, int64_t>(i + 1, i + 1);
   }
   return data;
 }
@@ -60,7 +60,7 @@ vector<tuple<int32_t, int32_t>> get_data_given_n(int n)
  * @param num_of_buckets The number of buckets to use.
  * @return The partition for the number.
  */
-int get_partition(int32_t n, int num_of_buckets)
+int get_partition(int64_t n, int num_of_buckets)
 {
   return n % num_of_buckets;
 }
@@ -93,8 +93,8 @@ int compute_input_chunk_size(int data_size, int num_of_threads)
  * @param buffers The buffers to move the item to.
  * @param num_of_buckets The number of buckets to use.
  */
-void move_element(array<atomic<int>, MAX_BUCKETS> &counter, const tuple<int32_t, int32_t> &item,
-                  vector<vector<tuple<int32_t, int32_t>>> &buffers, int num_of_buckets)
+void move_element(array<atomic<int>, MAX_BUCKETS> &counter, const tuple<int64_t, int64_t> &item,
+                  vector<vector<tuple<int64_t, int64_t>>> &buffers, int num_of_buckets)
 {
   int partition = get_partition(get<0>(item), num_of_buckets);
   int pos = increment_buffer_counter(counter, partition);
@@ -113,8 +113,8 @@ void move_element(array<atomic<int>, MAX_BUCKETS> &counter, const tuple<int32_t,
  * @param num_of_buckets The number of buckets to use.
  */
 void process_chunk(array<atomic<int>, MAX_BUCKETS> &counter, int thread_id, int start, int end,
-                   const vector<tuple<int32_t, int32_t>> &data,
-                   vector<vector<tuple<int32_t, int32_t>>> &buffers,
+                   const vector<tuple<int64_t, int64_t>> &data,
+                   vector<vector<tuple<int64_t, int64_t>>> &buffers,
                    int num_of_buckets)
 {
   // Set thread affinity to specific CPU core
@@ -138,7 +138,7 @@ void process_chunk(array<atomic<int>, MAX_BUCKETS> &counter, int thread_id, int 
  * @param num_of_buckets The number of buckets to use.
  * @param counter The counter to use.
  */
-void print_output(const vector<vector<tuple<int32_t, int32_t>>> &buffers, int num_of_buckets,
+void print_output(const vector<vector<tuple<int64_t, int64_t>>> &buffers, int num_of_buckets,
                   const array<atomic<int>, MAX_BUCKETS> &counter)
 {
   cout << "Data (first 10 elements from each partition): " << endl;
@@ -260,7 +260,7 @@ int main(int argc, char *argv[])
 
   vector<thread> threads;
   // Create output buffers for each partition
-  vector<vector<tuple<int32_t, int32_t>>> buffers(num_of_buckets, vector<tuple<int32_t, int32_t>>(data_size / num_of_buckets + 1));
+  vector<vector<tuple<int64_t, int64_t>>> buffers(num_of_buckets, vector<tuple<int64_t, int64_t>>(data_size / num_of_buckets + 1));
 
   auto start_time = high_resolution_clock::now(); // ------------------------ START TIME ------------------------
